+++ packages/cli/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @grackle-ai/cli
 
-This log was last generated on Mon, 23 Mar 2026 19:41:20 GMT and should not be manually modified.
+This log was last generated on Tue, 24 Mar 2026 04:08:14 GMT and should not be manually modified.
+
+## 0.74.0
+Tue, 24 Mar 2026 04:08:14 GMT
+
+### Minor changes
+
+- feat(powerline): scriptable stub runtime via JSON scenario in prompt
+
+### Updates
+
+- placeholder
+- placeholder
+- placeholder
+- placeholder
+- placeholder
 
 ## 0.73.1
 Mon, 23 Mar 2026 19:41:20 GMT